Moved the Event Model tests to their own file.
