+++ hivemind/modules/api/tests/test_signals.py
@@ -1,13 +1,10 @@
 """
 Tests for API Module Signals
 """
-import datetime
-from django.core.exceptions import ValidationError
 from django.test import TestCase
 from model_mommy import mommy
 
 from hivemind.modules.api.models import Profile
-from hivemind.modules.api.models.event import Event
 
 
 class TestSignals(TestCase):
@@ -34,35 +31,3 @@ class TestSignals(TestCase):
         self.assertEqual(profile.first_name, user.first_name)
         self.assertEqual(profile.last_name, user.last_name)
         self.assertEqual(profile.email, user.email)
-
-    def test_event_clean(self):
-        """
-        Test that when an event is created that
-        the dates are properly set and validated.
-        """
-
-        timezone = datetime.timezone(datetime.timedelta(0))
-        now = datetime.datetime.now(tz=timezone)
-
-        # End time before start time
-        with self.assertRaises(ValidationError):
-            Event(name='Test',
-                  description='Test',
-                  start_time=now,
-                  end_time=now - datetime.timedelta(days=1)
-                  ).clean()
-
-        # End time same as start time
-        with self.assertRaises(ValidationError):
-            Event(name='Test',
-                  description='Test',
-                  start_time=now,
-                  end_time=now
-                  ).clean()
-
-        # End time after start time
-        Event(name='Test',
-              description='Test',
-              start_time=now,
-              end_time=now + datetime.timedelta(days=1)
-              ).clean()
